LayoutTests/platform/glib/imported/w3c/web-platform-tests/loading/preloader-css-import-single-quote-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

FAIL Imported inline CSS is not blocked on pending CSS assert_true: The second resource start time should not be blocked on the first resource response expected true got false
